pyproject.toml
[tool.poetry]
name = "scan-receipt"
version = "0.1.0"
description = ""
authors = ["kazair <tokubd127@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit = "^1.43.2"
pydantic = "^2.11.0"
pydantic-settings = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
